refactor(lib/pokemon): resolve invalid import paths
